# apps/lab_backend/swagger-debug.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lab System API",
    "version": "1.0.0",
    "description": "API documentation for the Lab System backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/ratings": {
      "post": {
        "summary": "Submit a rating for an order",
        "tags": [
          "Ratings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rating submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or order already rated"
          },
          "403": {
            "description": "Unauthorized to rate this order"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "get": {
        "summary": "Get all ratings (For Lab Staff)",
        "tags": [
          "Ratings"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "rating",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            },
            "description": "Filter by specific rating score"
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by specific user"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Number of items per page for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all ratings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RatingWithDetails"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ratings/order/{orderId}": {
      "get": {
        "summary": "Get rating for a specific order",
        "tags": [
          "Ratings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The unique identifier of the order"
          }
        ],
        "responses": {
          "200": {
            "description": "Rating details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingWithDetails"
                }
              }
            }
          },
          "404": {
            "description": "No rating found for this order"
          }
        }
      }
    },
...
    "/api/tests": {
      "get": {
        "summary": "Returns the list of all lab tests",
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "clinic",
                "doctor",
                "patient"
              ]
            },
            "description": "User role to calculate specific discounts"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by test category"
          },
          {
            "in": "query",
            "name": "is_active",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          },
          {
            "in": "query",
            "name": "test_name",
            "schema": {
              "type": "string"
            },
            "description": "Search by partial test name"
          },
          {
            "in": "query",
            "name": "test_code",
            "schema": {
              "type": "string"
            },
            "description": "Search by partial test code"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Number of items per page for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "The list of tests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabTest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new lab test",
        "tags": [
          "Tests"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabTest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The test was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabTest"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/tests/{id}": {
      "get": {
        "summary": "Get a test by id",
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The test description by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabTest"
                }
              }
            }
          },
          "404": {
            "description": "Test not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing test",
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabTest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The test was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabTest"
                }
              }
            }
          },
          "404": {
            "description": "Test not found"
          }
        }
      },
      "delete": {
        "summary": "Soft delete a test",
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test deleted"
          },
          "404": {
            "description": "Test not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Rating": {
        "type": "object",
        "required": [
          "order_id",
          "user_id",
          "rating"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "order_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "remark": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RatingWithDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Rating"
          },
          {
            "type": "object",
            "properties": {
              "user_name": {
                "type": "string"
              },
              "user_email": {
                "type": "string"
              },
              "user_phone": {
                "type": "string"
              },
              "patient_name": {
                "type": "string"
              },
              "patient_age": {
                "type": "integer"
              },
              "order_status": {
                "type": "string"
              },
              "priority": {
                "type": "string"
              },
              "final_price_mmk": {
                "type": "number"
              },
              "order_created_at": {
                "type": "string",
                "format": "date-time"
              },
              "order_tests": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "test_name": {
                      "type": "string"
                    },
                    "test_code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "LabTest": {
        "type": "object",
        "required": [
          "test_name",
          "test_code",
          "base_price_mmk"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "test_name": {
            "type": "string"
          },
          "test_code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "base_price_mmk": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}